+++ houses/views.py
@@ -1,4 +1,5 @@
 import datetime
+import sys
 
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
@@ -27,8 +28,9 @@ def house_create(request):
 	if request.method == 'POST':
 		house = House()
 		house.user = request.user
-		if request.POST['street_number'] and request.POST['street_number'] and request.POST['street_name'] and request.POST['city'] and request.POST['prov_state'] and request.POST['country']:
-			print(request.POST)
+		if ('street_number' or 'street_number' or 'street_name' or 'city' or 'prov_state' or 'country' or 'place_id' or 'lat' or 'lon') not in request.POST:
+			return render(request, 'houses/house_create.html', {'error': 'There is an issue with the address inputted!', 'GOOGLE_API_KEY': GOOGLE_API_KEY, 'captcha': captcha})
+		else:
 			# Create House
 			house.street_number = request.POST['street_number']
 			house.street_name = request.POST['street_name']
@@ -49,14 +51,21 @@ def house_create(request):
 			init_bill_set.year = date.year
 			init_bill_set.house = house
 			init_bill_set.save()
-
-			# Load walk score for house
-			house.load_walk_score()
-			# Load house image
-			house_image = load_house_image(house)
+			
+			# TODO: Put these in a background task or make them async as they take 90% of the request time
+			# TODO: This can be remove at a later time, but at the moment it costs too much to test it constantly
+			if 'test' in sys.argv:
+				pass
+			else:
+				house.load_walk_score()
+				house_image = load_house_image(house)
 
 			# Create GarbageDay for house if info present
-			if request.POST['LastGarbageDay'] and request.POST['NextGarbageDay']:
+			if ('LastGarbageDay' or 'NextGarbageDay') not in request.POST:
+				pass
+			elif (request.POST['LastGarbageDay'] or request.POST['NextGarbageDay']) is '':
+				pass
+			else:
 				garbageday = GarbageDay()
 				garbageday.house = house
 				garbageday.user = request.user
@@ -65,7 +74,6 @@ def house_create(request):
 				garbageday.save()
 
 			return redirect('house_detail', pk=house.id)
-		return render(request, 'houses/house_create.html', {'error': 'There is an issue with the address inputted!', 'GOOGLE_API_KEY': GOOGLE_API_KEY, 'captcha': captcha})
 	else:
 		return render(request, 'houses/house_create.html', {'GOOGLE_API_KEY': GOOGLE_API_KEY, 'captcha': captcha})
 
@@ -79,7 +87,13 @@ def house_bill_add(request, pk):
 	if request.method == 'POST':
 		bill = Bill()
 		bill.user = request.user
-		if request.POST['type'] == '' or request.POST['date'] == '' or request.POST['amount'] == '':
+		if 'type' not in request.POST:
+			return render(request, 'houses/house_bill_add.html', {'house': house, 'error': 'You have entered invalid data!'})
+
+		if 'date' not in request.POST:
+			return render(request, 'houses/house_bill_add.html', {'house': house, 'error': 'You have entered invalid data!'})
+
+		if 'amount' not in request.POST:
 			return render(request, 'houses/house_bill_add.html', {'house': house, 'error': 'You have entered invalid data!'})
 
 		bill.type = request.POST['type']
@@ -127,7 +141,9 @@ def house_invite(request, pk):
 		raise Http404
 	captcha = Captcha()
 	if request.method == 'POST':
-		if request.POST['email'] != '':
+		if 'email' not in request.POST:
+			return render(request, 'houses/house_invite.html', {'house': house, 'captcha': captcha, 'error': 'Please enter a valid email address'})
+		else:
 			invitation = Invitation()
 			invitation.house = house
 			invitation.sender = request.user
@@ -138,14 +154,15 @@ def house_invite(request, pk):
 		return redirect('house_detail', pk=pk)
 	return render(request, 'houses/house_invite.html', {'house': house, 'captcha': captcha})
 
-
+# PK is for house and ID is for invitation
 @login_required(login_url="account_login")
 def house_invite_remove(request, pk, id):
 	house = get_object_or_404(House, pk=pk)
 	if request.user.id != house.user.id:
 		raise Http404
-	invite = get_object_or_404(Invitation, id=id)
-	invite.delete()
+	if request.method == 'POST':
+		invite = get_object_or_404(Invitation, id=id)
+		invite.delete()
 	return redirect('house_detail', pk=pk)
 
 
@@ -156,8 +173,9 @@ def house_invite_accept(request, pk, id):
 	if request.user.email != invite.target:
 		raise Http404
 
-	house.members.add(request.user)
-	invite.delete()
+	if request.method == 'POST':
+		house.members.add(request.user)
+		invite.delete()
 	return redirect('house_detail', pk=pk)
 
 
@@ -166,17 +184,19 @@ def house_invite_decline(request, pk, id):
 	invite = get_object_or_404(Invitation, id=id)
 	if request.user.email != invite.target:
 		raise Http404
-	invite.delete()
+	if request.method == 'POST':
+		invite.delete()
 	return redirect('main_dashboard')
 
-
+# pk is for the house, and ID is for the userid of the user you want to remove
 @login_required(login_url="account_login")
 def house_member_remove(request, pk, id):
 	house = get_object_or_404(House, pk=pk)
 	if request.user.id != house.user.id:
 		raise Http404
-	member = house.members.filter(id=id).first()
-	house.members.remove(member)
+	if request.method == 'POST':
+		member = house.members.filter(id=id).first()
+		house.members.remove(member)
 	return redirect('house_detail', pk=pk)
 
 
@@ -188,10 +208,21 @@ def house_add_room(request, pk):
 	if request.method == 'POST':
 		room = Room()
 		room.user = request.user
-		room.name = request.POST['name']
 		room.house = house
-		room.price = request.POST['price']
-		room.description = request.POST['description']
+		if 'name' not in request.POST:
+			return render(request, 'houses/room_add.html', {'house': house, 'error': 'Please make sure to fill in all required details'})
+		else:
+			room.name = request.POST['name']
+
+		if 'price' not in request.POST:
+			return render(request, 'houses/room_add.html', {'house': house, 'error': 'Please make sure to fill in all required details'})
+		else:
+			room.price = request.POST['price']
+
+		if 'description' not in request.POST:
+			return render(request, 'houses/room_add.html', {'house': house, 'error': 'Please make sure to fill in all required details'})
+		else:
+			room.description = request.POST['description']
 		room.save()
 		try:
 			for file in request.FILES.getlist('images'):
